change argument order in depth kernel
